+++ src/Mods/AstralArena/ScriptExtender/Lua/AstralArena/Shared/Constants.lua
@@ -1,6 +1,6 @@
 return {
     ModuleUUID = "29c48c80-8777-f7b5-6bb8-376c1c5d8db6",
-    DisplayVersion = "0.3.1-alpha.2",
+    DisplayVersion = "0.3.1-alpha.3",
     ArenaLevel = "AA_Arena_Main",
     SchemaVersion = 1,
     EntrantCount = 8,